Rename api/standard package to semconv
